telegram_bot/requirements.txt
# Telegram Bot Dependencies
python-telegram-bot[ext]>=20.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
toml>=0.10.0